refactor(content_ui): consolidate page graphql
