+++ frontend/plugins/content_ui/src/modules/cms/pages/graphql/index.ts
@@ -1,1 +1,2 @@
+export * from './mutations';
 export * from './queries';
