+++ frontend/plugins/content_ui/src/modules/cms/pages/graphql/mutations/index.ts
@@ -1,0 +1,1 @@
+export { PAGES_ADD, PAGES_EDIT, PAGES_REMOVE } from './pagesMutations';
