+++ frontend/plugins/content_ui/src/modules/cms/pages/graphql/queries/pagesListQueries.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const PAGE_LIST = gql`
-  query PageList(
+  query cmsPageList(
     $clientPortalId: String!
     $searchValue: String
     $language: String
