+++ frontend/plugins/content_ui/src/modules/cms/pages/hooks/useAddPage.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@apollo/client';
-import { PAGES_ADD } from '../graphql/mutations/pagesMutations';
+import { PAGES_ADD } from '@/cms/pages/graphql';
 
 export const useAddPage = () => {
   const [addPageMutation, { loading }] = useMutation(PAGES_ADD);
