+++ frontend/plugins/content_ui/src/modules/cms/pages/hooks/usePageDetail.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@apollo/client';
-import { PAGE_DETAIL } from '../graphql/queries/pageDetailQuery';
+import { PAGE_DETAIL } from '@/cms/pages/graphql';
 import { IPage } from '../types/pageTypes';
 
 export const usePageDetail = (id: string) => {
